+++ docs/source/whatsnew.rst
@@ -7,6 +7,10 @@ What's New
 
     v4.0.0 is a significant rewrite of a lot of the methods and API calls to reduce the number of API calls needed in large Orgs with a many Users/Locations/Workspaces. I have tried to document all the breaking changes, but there may be some I haven't realized yet. The v4.0.0 series is changing rapidly as I find them.
 
+v4.2.3
+------
+- Fixed `spark_id` for :class:`AutoAttendant` and :class:`HuntGroup`
+
 v4.2.2
 ------
 - Added `uuid` to :meth:`HuntGroupList.get()` to support UUID lookup for CDR correlation